Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-    int num1,num2,product;
+    float num1,num2,product;
     scanf("%f%f", &num1,&num2);
     product= num1*num2;
     printf("Product: %.2f",product);
